+++ app/src/main/java/com/easyhz/patchnote/ui/screen/defect/defectDetail/contract/DetailSideEffect.kt
@@ -1,0 +1,9 @@
+package com.easyhz.patchnote.ui.screen.defect.defectDetail.contract
+
+import com.easyhz.patchnote.core.common.base.UiSideEffect
+import com.easyhz.patchnote.core.model.defect.DefectMainItem
+
+sealed class DetailSideEffect: UiSideEffect() {
+    data object NavigateToUp: DetailSideEffect()
+    data class NavigateToDefectCompletion(val defectMainItem: DefectMainItem): DetailSideEffect()
+}
