--- app/src/main/java/com/easyhz/patchnote/ui/screen/defectDetail/contract/DetailSideEffect.kt
@@ -1,9 +1,0 @@
-package com.easyhz.patchnote.ui.screen.defectDetail.contract
-
-import com.easyhz.patchnote.core.common.base.UiSideEffect
-import com.easyhz.patchnote.core.model.defect.DefectMainItem
-
-sealed class DetailSideEffect: UiSideEffect() {
-    data object NavigateToUp: DetailSideEffect()
-    data class NavigateToDefectCompletion(val defectMainItem: DefectMainItem): DetailSideEffect()
-}
